requirements.txt
opencv-python
numpy
pandas
matplotlib
scipy
scikit-image
Pillow
openpyxl
scikit-learn
